chore: VsTest test classes refactoring
