+++ src/Stryker.Core/Stryker.Core.UnitTest/TestBase.cs
@@ -1,4 +1,3 @@
-using System;
 using Microsoft.Extensions.Logging;
 using Stryker.Core.Logging;
 
@@ -6,10 +5,8 @@ namespace Stryker.Core.UnitTest
 {
     public abstract class TestBase
     {
-        protected TestBase()
-        {
+        protected TestBase() =>
             // initialize loggerfactory to prevent exceptions
             ApplicationLogging.LoggerFactory = new LoggerFactory();
-        }
     }
 }
