+++ src/Stryker.Core/Stryker.Core.UnitTest/TestRunners/VsTestRunnerPoolTests.cs
@@ -1,35 +1,13 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.IO.Abstractions.TestingHelpers;
 using System.Linq;
-using System.Reflection;
-using System.Threading;
-using System.Threading.Tasks;
-using Buildalyzer;
-using Microsoft.Extensions.Logging.Abstractions;
-using Microsoft.TestPlatform.VsTestConsole.TranslationLayer.Interfaces;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel.Client;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel.Client.Interfaces;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel.DataCollection;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel.DataCollector.InProcDataCollector;
 using Moq;
 using Shouldly;
 using Stryker.Core.CoverageAnalysis;
 using Stryker.Core.Exceptions;
-using Stryker.Core.Initialisation;
-using Stryker.Core.MutantFilters;
 using Stryker.Core.Mutants;
-using Stryker.Core.MutationTest;
 using Stryker.Core.Options;
-using Stryker.Core.ProjectComponents;
-using Stryker.Core.Reporters;
-using Stryker.Core.TestRunners;
-using Stryker.Core.TestRunners.VsTest;
-using Stryker.Core.ToolHelpers;
-using Stryker.DataCollector;
 using Xunit;
 
 namespace Stryker.Core.UnitTest.TestRunners
@@ -38,397 +16,32 @@ namespace Stryker.Core.UnitTest.TestRunners
     /// This class hosts the VsTestRunner related tests. The design of VsTest implies the creation of many mocking objects, so the tests may be hard to read.
     /// This is sad but expected. Please use caution when changing/creating tests.
     /// </summary>
-    public class VsTestRunnerPoolTests : TestBase
+    public class VsTestRunnerPoolTests : VsTestMockingHelper
     {
-        private readonly string _testAssemblyPath;
-        private readonly ProjectInfo _targetProject;
-        private readonly MockFileSystem _fileSystem;
-        private readonly Mutant _mutant;
-        private readonly Mutant _otherMutant;
-        private readonly CsharpFolderComposite _projectContents;
-        private readonly Uri _executorUri;
-        private readonly TestProperty _coverageProperty;
-        private readonly TestProperty _unexpectedCoverageProperty;
         private readonly Uri _xUnitUri;
-        private readonly TimeSpan _testDefaultDuration = TimeSpan.FromSeconds(1);
 
         // initialize the test context and mock objects
         public VsTestRunnerPoolTests()
         {
-            var currentDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            var filesystemRoot = Path.GetPathRoot(currentDirectory);
-
-            var sourceFile = File.ReadAllText(currentDirectory + "/TestResources/ExampleSourceFile.cs");
-            var testProjectPath = FilePathUtils.NormalizePathSeparators(Path.Combine(filesystemRoot!, "TestProject", "TestProject.csproj"));
-            var projectUnderTestPath = FilePathUtils.NormalizePathSeparators(Path.Combine(filesystemRoot, "ExampleProject", "ExampleProject.csproj"));
-            const string defaultTestProjectFileContents = @"<Project Sdk=""Microsoft.NET.Sdk"">
-    <PropertyGroup>
-        <TargetFramework>netcoreapp2.0</TargetFramework>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include=""Microsoft.NET.Test.Sdk"" Version = ""15.5.0"" />
-        <PackageReference Include=""xunit"" Version=""2.3.1"" />
-        <PackageReference Include=""xunit.runner.visualstudio"" Version=""2.3.1"" />
-        <DotNetCliToolReference Include=""dotnet-xunit"" Version=""2.3.1"" />
-    </ItemGroup>
-    <ItemGroup>
-        <ProjectReference Include=""..\ExampleProject\ExampleProject.csproj"" />
-    </ItemGroup>
-</Project>";
-            _testAssemblyPath = FilePathUtils.NormalizePathSeparators(Path.Combine(filesystemRoot, "_firstTest", "bin", "Debug", "TestApp.dll"));
-            _executorUri = new Uri("exec://nunit");
             _xUnitUri = new Uri("executor://xunit/VsTestRunner2/netcoreapp");
-            var firstTest = BuildCase("T0");
-            var secondTest = BuildCase("T1");
-
-            var content = new CsharpFolderComposite();
-            _fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
-            {
-                { projectUnderTestPath, new MockFileData(defaultTestProjectFileContents)},
-                { Path.Combine(filesystemRoot, "ExampleProject", "Recursive.cs"), new MockFileData(sourceFile)},
-                { Path.Combine(filesystemRoot, "ExampleProject", "OneFolderDeeper", "Recursive.cs"), new MockFileData(sourceFile)},
-                { testProjectPath, new MockFileData(defaultTestProjectFileContents)},
-                { _testAssemblyPath!, new MockFileData("Bytecode") },
-                { Path.Combine(filesystemRoot, "app", "bin", "Debug", "AppToTest.dll"), new MockFileData("Bytecode") },
-            });
-            _coverageProperty = TestProperty.Register(CoverageCollector.PropertyName, CoverageCollector.PropertyName, typeof(string), typeof(TestResult));
-            _unexpectedCoverageProperty = TestProperty.Register(CoverageCollector.OutOfTestsPropertyName, CoverageCollector.OutOfTestsPropertyName, typeof(string), typeof(TestResult));
-            _mutant = new Mutant { Id = 0 };
-            _otherMutant = new Mutant { Id = 1 };
-            _projectContents = content;
-            _projectContents.Add(new CsharpFileLeaf { Mutants = new[] { _otherMutant, _mutant } });
-            _targetProject = new ProjectInfo(_fileSystem)
-            {
-                TestProjectAnalyzerResults = new List<IAnalyzerResult> {
-                    TestHelper.SetupProjectAnalyzerResult(
-                    properties: new Dictionary<string, string>() {
-                        { "TargetDir", Path.GetDirectoryName(_testAssemblyPath) },
-                        { "TargetFileName", Path.GetFileName(_testAssemblyPath) }
-                    },
-                    targetFramework: "netcoreapp2.1").Object
-                },
-                ProjectUnderTestAnalyzerResult = TestHelper.SetupProjectAnalyzerResult(
-                    properties: new Dictionary<string, string>() {
-                        { "TargetDir", Path.Combine(filesystemRoot, "app", "bin", "Debug") },
-                        { "TargetFileName", "AppToTest.dll" },
-                        { "Language", "C#" }
-                    }).Object,
-                ProjectContents = _projectContents
-            };
-
-            TestCases = new List<TestCase> { firstTest, secondTest };
         }
 
-        private List<TestCase> TestCases { get; }
-
         // simulate the discovery of tests
-        private static void DiscoverTests(ITestDiscoveryEventsHandler discoveryEventsHandler, ICollection<TestCase> tests, bool aborted) =>
-            Task.Run(() => discoveryEventsHandler.HandleDiscoveredTests(tests)).
-                ContinueWith((_, u) => discoveryEventsHandler.HandleDiscoveryComplete((int)u, null, aborted), tests.Count);
-
-        private TestCase BuildCase(string name) => new(name, _executorUri, _testAssemblyPath) { Id = new Guid() };
-
-        private TestCase FindOrBuildCase(string testResultId) => TestCases.FirstOrDefault(t => t.FullyQualifiedName == testResultId) ?? BuildCase(testResultId);
 
         // mock a VsTest run. Provides test result one by one at 10 ms intervals
         // note: a lot of information is still missing (vs real VsTest). You will have to add them if your test requires them
-        private void MoqTestRun(ITestRunEventsHandler testRunEvents, IReadOnlyList<TestResult> testResults,
-            TestCase timeOutTest = null) =>
-            Task.Run(() =>
-            {
-                var timer = new Stopwatch();
-                testRunEvents.HandleTestRunStatsChange(
-                    new TestRunChangedEventArgs(new TestRunStatistics(0, null), null, timeOutTest == null ? null : new[] { timeOutTest }));
-
-                for (var i = 0; i < testResults.Count; i++)
-                {
-                    Thread.Sleep(10);
-                    testResults[i].EndTime = DateTimeOffset.Now;
-                    testRunEvents.HandleTestRunStatsChange(new TestRunChangedEventArgs(
-                        new TestRunStatistics(i + 1, null), new[] { testResults[i] }, null));
-                }
-
-                if (timeOutTest != null)
-                {
-                    testRunEvents.HandleTestRunStatsChange(new TestRunChangedEventArgs(
-                        new TestRunStatistics(testResults.Count, null), null, new[] { timeOutTest }));
-                }
-
-                Thread.Sleep(10);
-                testRunEvents.HandleTestRunComplete(
-                    new TestRunCompleteEventArgs(new TestRunStatistics(testResults.Count, null), false, false, null,
-                        null, timer.Elapsed),
-                    new TestRunChangedEventArgs(null, Array.Empty<TestResult>(), new List<TestCase>()),
-                    null,
-                    null);
-            });
 
         // setup a simple (mock) test run where all tests fail or succeed
-        private void SetupMockTestRun(Mock<IVsTestConsoleWrapper> mockVsTest, bool testResult, EventWaitHandle synchroObject, ICollection<TestCase> testCases)
-        {
-            var results = new List<(string, bool)>(testCases.Count);
-            results.AddRange(testCases.Select(t => (t.FullyQualifiedName, testResult)));
-            SetupMockTestRun(mockVsTest, results, synchroObject);
-        }
-
-        private void SetupMockTestRun(Mock<IVsTestConsoleWrapper> mockVsTest, IEnumerable<(string id, bool success)> testResults,
-            EventWaitHandle synchroObject, ICollection<TestCase> testCases = null)
-        {
-            var results = new List<TestResult>();
-            testCases ??= TestCases;
-            foreach (var (testResultId, success) in testResults)
-            {
-                var testCase = testCases.FirstOrDefault(t => t.FullyQualifiedName == testResultId) ?? BuildCase(testResultId);
-                results.Add(new TestResult(testCase)
-                {
-                    Outcome = success ? TestOutcome.Passed : TestOutcome.Failed,
-                    ComputerName = ".",
-                    Duration = _testDefaultDuration
-                });
-            }
-            mockVsTest.Setup(x =>
-                x.RunTestsWithCustomTestHostAsync(
-                    It.Is<IEnumerable<string>>(t => t.Any(source => source == _testAssemblyPath)),
-                    It.Is<string>(settings => !settings.Contains("<Coverage")),
-                    It.Is<TestPlatformOptions>(o => o != null && o.TestCaseFilter == null),
-                    It.IsAny<ITestRunEventsHandler>(),
-                    It.IsAny<ITestHostLauncher>())).Callback(
-                (IEnumerable<string> _, string _, TestPlatformOptions _, ITestRunEventsHandler testRunEvents,
-                    ITestHostLauncher _) =>
-                {
-                    // generate test results
-                    MoqTestRun(testRunEvents, results);
-                    synchroObject.Set();
-                }).Returns(Task.CompletedTask);
-        }
 
         // setup a customized coverage capture run, using provided coverage results
-        private void SetupMockCoverageRun(Mock<IVsTestConsoleWrapper> mockVsTest, IReadOnlyDictionary<string, string> coverageResults, EventWaitHandle endProcess) =>
-            mockVsTest.Setup(x =>
-                x.RunTestsWithCustomTestHostAsync(
-                    It.Is<IEnumerable<string>>(t => t.Any(source => source == _testAssemblyPath)),
-                    It.Is<string>(settings => settings.Contains("<Coverage")),
-                    It.Is<TestPlatformOptions>(o => o != null && o.TestCaseFilter == null),
-                    It.IsAny<ITestRunEventsHandler>(),
-                    It.IsAny<ITestHostLauncher>())).Callback(
-                (IEnumerable<string> _, string _, TestPlatformOptions _, ITestRunEventsHandler testRunEvents,
-                    ITestHostLauncher _) =>
-                {
-                    // generate test results
-                    var results = new List<TestResult>(coverageResults.Count);
-                    foreach (var (key, value) in coverageResults)
-                    {
-                        var result = new TestResult(FindOrBuildCase(key))
-                        {
-                            DisplayName = key,
-                            Outcome = TestOutcome.Passed,
-                            ComputerName = "."
-                        };
-                        if (value != null)
-                        {
-                            var coveredList = value.Split('|');
-                            result.SetPropertyValue(_coverageProperty, coveredList[0]);
-                            if (coveredList.Length > 1)
-                            {
-                                result.SetPropertyValue(_unexpectedCoverageProperty, coveredList[1]);
-                            }
-                        }
-
-                        results.Add(result);
-                    }
-                    MoqTestRun(testRunEvents, results);
-                    endProcess.Set();
-                }).Returns(Task.CompletedTask);
-
-        private void SetupMockCoveragePerTestRunP(Mock<IVsTestConsoleWrapper> mockVsTest, IReadOnlyDictionary<string, string> coverageResults, EventWaitHandle endProcess) =>
-            mockVsTest.Setup(x =>
-                x.RunTestsWithCustomTestHostAsync(
-                    It.Is<IEnumerable<TestCase>>(t => t.Any()),
-                    It.Is<string>(settings => settings.Contains("<Coverage")),
-                    It.Is<TestPlatformOptions>(o => o != null && o.TestCaseFilter == null),
-                    It.IsAny<ITestRunEventsHandler>(),
-                    It.IsAny<ITestHostLauncher>())).Callback(
-                (IEnumerable<TestCase> testCases, string _, TestPlatformOptions _, ITestRunEventsHandler testRunEvents,
-                    ITestHostLauncher _) =>
-                {
-                    // generate test results
-                    var results = new List<TestResult>(coverageResults.Count);
-                    foreach (var (key, value) in coverageResults)
-                    {
-                        var coveredList = value.Split('|');
-                        if (!testCases.Any(t => t.DisplayName== key))
-                        {
-                            continue;
-                        }
-                        var result = new TestResult(FindOrBuildCase(key))
-                        {
-                            DisplayName = key,
-                            Outcome = TestOutcome.Passed,
-                            ComputerName = "."
-                        };
-                        result.SetPropertyValue(_coverageProperty, coveredList[0]);
-                        if (coveredList.Length > 1)
-                        {
-                            result.SetPropertyValue(_unexpectedCoverageProperty, coveredList[1]);
-                        }
-                        results.Add(result);
-                    }
-                    MoqTestRun(testRunEvents, results);
-                    endProcess.Set();
-                }).Returns(Task.CompletedTask);
 
         // setup a customized partial test runs, using provided test results
-        private void SetupMockPartialTestRun(Mock<IVsTestConsoleWrapper> mockVsTest, IReadOnlyDictionary<string, string> results, EventWaitHandle endProcess) =>
-            mockVsTest.Setup(x =>
-                x.RunTestsWithCustomTestHostAsync(
-                    It.IsAny<IEnumerable<TestCase>>(),
-                    It.Is<string>(s => !s.Contains("<Coverage")),
-                    It.Is<TestPlatformOptions>(o => o != null && o.TestCaseFilter == null),
-                    It.IsAny<ITestRunEventsHandler>(),
-                    It.IsAny<ITestHostLauncher>())).Callback(
-                (IEnumerable<TestCase> sources, string settings, TestPlatformOptions _, ITestRunEventsHandler testRunEvents,
-                    ITestHostLauncher _) =>
-                {
-                    var collector = new CoverageCollector();
-                    var start = new TestSessionStartArgs
-                    {
-                        Configuration = settings
-                    };
-                    var mock = new Mock<IDataCollectionSink>(MockBehavior.Loose);
-                    collector.Initialize(mock.Object);
-                    collector.TestSessionStart(start);
-
-                    var mutants = collector.MutantList;
-                    if (!results.ContainsKey(mutants))
-                    {
-                        throw new ArgumentException($"Unexpected mutant run {mutants}");
-                    }
-
-                    var tests = sources.ToList();
-                    var data = results[mutants].Split(',').Select(e => e.Split('=')).ToList();
-                    if (data.Count != tests.Count)
-                    {
-                        throw new ArgumentException($"Invalid number of tests for mutant run {mutants}: found {tests.Count}, expected {data.Count}");
-                    }
-
-                    var runResults = new List<TestResult>(data.Count);
-                    foreach (var strings in data)
-                    {
-                        var matchingTest = tests.FirstOrDefault(t => t.FullyQualifiedName == strings[0]);
-                        if (matchingTest == null)
-                        {
-                            throw new ArgumentException($"Test {strings[0]} not run for mutant {mutants}.");
-                        }
-
-                        var result = new TestResult(matchingTest)
-                            { Outcome = strings[1] == "F" ? TestOutcome.Failed : TestOutcome.Passed, ComputerName = "." };
-                        runResults.Add(result);
-                    }
-                    // setup a normal test run
-                    MoqTestRun(testRunEvents, runResults);
-                    collector.TestSessionEnd(new TestSessionEndArgs());
-
-                    endProcess.Set();
-                }).Returns(Task.CompletedTask);
-
-        private void SetupMockTimeOutTestRun(Mock<IVsTestConsoleWrapper> mockVsTest, IReadOnlyDictionary<string, string> results, string timeoutTest, EventWaitHandle endProcess) =>
-            mockVsTest.Setup(x =>
-                x.RunTestsWithCustomTestHostAsync(
-                    It.IsAny<IEnumerable<TestCase>>(),
-                    It.IsAny<string>(),
-                    It.Is<TestPlatformOptions>(o => o != null && o.TestCaseFilter == null),
-                    It.IsAny<ITestRunEventsHandler>(),
-                    It.IsAny<ITestHostLauncher>())).Callback(
-                (IEnumerable<TestCase> sources, string settings, TestPlatformOptions _, ITestRunEventsHandler testRunEvents,
-                    ITestHostLauncher _) =>
-                {
-                    var collector = new CoverageCollector();
-                    var start = new TestSessionStartArgs
-                    {
-                        Configuration = settings
-                    };
-                    var mock = new Mock<IDataCollectionSink>(MockBehavior.Loose);
-                    TestCase timeOutTestCase = null;
-                    collector.Initialize(mock.Object);
-                    collector.TestSessionStart(start);
-
-                    var mutants = collector.MutantList;
-                    if (!results.ContainsKey(mutants))
-                    {
-                        throw new ArgumentException($"Unexpected mutant run {mutants}");
-                    }
-
-                    var tests = sources.ToList();
-                    var data = results[mutants].Split(',').Select(e => e.Split('=')).ToList();
-                    if (data.Count != tests.Count)
-                    {
-                        throw new ArgumentException($"Invalid number of tests for mutant run {mutants}: found {tests.Count}, expected {data.Count}");
-                    }
-
-                    var runResults = new List<TestResult>(data.Count);
-                    foreach (var strings in data)
-                    {
-                        var matchingTest = tests.FirstOrDefault(t => t.FullyQualifiedName == strings[0]);
-                        if (matchingTest == null)
-                        {
-                            throw new ArgumentException($"Test {strings[0]} not run for mutant {mutants}.");
-                        }
-                        if (matchingTest.FullyQualifiedName == timeoutTest)
-                        {
-                            timeOutTestCase = matchingTest;
-                        }
-                        var result = new TestResult(matchingTest)
-                            { Outcome = strings[1] == "F" ? TestOutcome.Failed : TestOutcome.Passed, ComputerName = "." };
-                        runResults.Add(result);
-                    }
-                    // setup a normal test run
-                    MoqTestRun(testRunEvents, runResults, timeOutTestCase);
-                    collector.TestSessionEnd(new TestSessionEndArgs());
-
-                    endProcess.Set();
-                }).Returns(Task.CompletedTask);
-
-        private Mock<IVsTestConsoleWrapper> BuildVsTestRunnerPool(StrykerOptions options,
-            WaitHandle endProcess,
-            out VsTestRunnerPool runner, ICollection<TestCase> testCases=null,
-            bool succeed = true)
-        {
-            testCases ??= TestCases;
-            var mockedVsTestConsole = new Mock<IVsTestConsoleWrapper>(MockBehavior.Strict);
-            mockedVsTestConsole.Setup(x => x.StartSession());
-            mockedVsTestConsole.Setup(x => x.InitializeExtensions(It.IsAny<List<string>>()));
-            mockedVsTestConsole.Setup(x => x.AbortTestRun());
-            mockedVsTestConsole.Setup(x => x.EndSession());
-            mockedVsTestConsole.Setup(x =>
-                x.DiscoverTests(It.Is<IEnumerable<string>>(d => d.Any(e => e == _testAssemblyPath)),
-                    It.IsAny<string>(),
-                    It.IsAny<ITestDiscoveryEventsHandler>())).Callback(
-                (IEnumerable<string> _, string _, ITestDiscoveryEventsHandler discoveryEventsHandler) =>
-                    DiscoverTests(discoveryEventsHandler, testCases, false));
-
-            var context = new VsTestContextInformation(
-                options,
-                _targetProject,
-                new Mock<IVsTestHelper>().Object,
-                _fileSystem,
-                _=> mockedVsTestConsole.Object,
-                hostBuilder: _ => new MoqHost(endProcess, succeed, false),
-                NullLogger.Instance
-            );
-            context.Initialize();
-            runner = new VsTestRunnerPool(context,
-                NullLogger.Instance,
-                (information, _) => new VsTestRunner(information, 0, NullLogger.Instance));
-            return mockedVsTestConsole;
-        }
 
         [Fact]
         public void RunInitialTestsOn()
         {
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), endProcess, out var runner);
-            SetupMockTestRun(mockVsTest, new[] { ("T0", false), ("T1", true) }, endProcess);
+            var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), out var runner);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", false), ("T1", true) });
             var result = runner.InitialTest();
             // tests are successful => run should be successful
             result.FailingTests.Count.ShouldBe(1);
@@ -437,9 +50,8 @@ namespace Stryker.Core.UnitTest.TestRunners
         [Fact]
         public void HandleVsTestCreationFailure()
         {
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), endProcess, out var runner, TestCases, false);
-            SetupMockTestRun(mockVsTest, new[] { ("T0", false), ("T1", true) }, endProcess);
+            var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), out var runner,  succeed:false);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", false), ("T1", true) });
 
             Action action = () => runner.InitialTest();
             action.ShouldThrow<GeneralStrykerException>();
@@ -448,10 +60,9 @@ namespace Stryker.Core.UnitTest.TestRunners
         [Fact]
         public void RunTests()
         {
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), endProcess, out var runner);
-            SetupMockTestRun(mockVsTest, true, endProcess, TestCases);
-            var result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            var mockVsTest = BuildVsTestRunnerPool(new StrykerOptions(), out var runner);
+            SetupMockTestRun(mockVsTest, true, TestCases);
+            var result = runner.TestMultipleMutants(null, new []{Mutant}, null);
             // tests are successful => run should be successful
             result.FailingTests.IsEmpty.ShouldBeTrue();
         }
@@ -460,14 +71,11 @@ namespace Stryker.Core.UnitTest.TestRunners
         public void DetectTestsErrors()
         {
             var options = new StrykerOptions();
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.AutoReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
-                SetupMockTestRun(mockVsTest, false, endProcess, TestCases);
-                var result = runner.TestMultipleMutants(null, new []{_mutant}, null);
-                // run is failed
-                result.FailingTests.IsEmpty.ShouldBeFalse();
-            }
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
+            SetupMockTestRun(mockVsTest, false, TestCases);
+            var result = runner.TestMultipleMutants(null, new []{Mutant}, null);
+            // run is failed
+            result.FailingTests.IsEmpty.ShouldBeFalse();
         }
 
         [Fact]
@@ -478,18 +86,15 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.AutoReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
 
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = "1;" }, endProcess);
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
-                SetupMockTimeOutTestRun(mockVsTest, new Dictionary<string, string> { ["0"] = "T0=S;T1=S" }, "T0", endProcess);
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = "1;" });
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
+            SetupMockTimeOutTestRun(mockVsTest, new Dictionary<string, string> { ["0"] = "T0=S;T1=S" }, "T0");
 
-                var result = runner.TestMultipleMutants(null, new []{_mutant}, null);
-                result.TimedOutTests.IsEmpty.ShouldBeFalse();
-            }
+            var result = runner.TestMultipleMutants(null, new[] { Mutant }, null);
+            result.TimedOutTests.IsEmpty.ShouldBeFalse();
         }
 
         [Fact]
@@ -497,19 +102,16 @@ namespace Stryker.Core.UnitTest.TestRunners
         {
             var options = new StrykerOptions();
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
 
-                mockVsTest.Setup(x => x.CancelTestRun()).Verifiable();
-                SetupMockTestRun(mockVsTest, false, endProcess, TestCases);
+            mockVsTest.Setup(x => x.CancelTestRun()).Verifiable();
+            SetupMockTestRun(mockVsTest, false, TestCases);
 
-                var result = runner.TestMultipleMutants(null, new []{_mutant},  ((_, _, _, _) => false));
-                // verify Abort has been called
-                Mock.Verify(mockVsTest);
-                // and test run is failed
-                result.FailingTests.IsEmpty.ShouldBeFalse();
-            }
+            var result = runner.TestMultipleMutants(null, new[] { Mutant }, ((_, _, _, _) => false));
+            // verify Abort has been called
+            Mock.Verify(mockVsTest);
+            // and test run is failed
+            result.FailingTests.IsEmpty.ShouldBeFalse();
         }
 
         [Fact]
@@ -520,19 +122,16 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.SkipUncoveredMutants
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
-                // test 0 and 1 cover mutant 1
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = "0;" }, endProcess);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
+            // test 0 and 1 cover mutant 1
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = "0;" });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
-                // one mutant is covered by tests 0 and 1
-                _mutant.CoveringTests.IsEmpty.ShouldBe(false);
-                _otherMutant.CoveringTests.IsEmpty.ShouldBe(true);
-                _otherMutant.ResultStatus.ShouldBe(MutantStatus.NoCoverage);
-            }
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
+            // one mutant is covered by tests 0 and 1
+            Mutant.CoveringTests.IsEmpty.ShouldBe(false);
+            OtherMutant.CoveringTests.IsEmpty.ShouldBe(true);
+            OtherMutant.ResultStatus.ShouldBe(MutantStatus.NoCoverage);
         }
 
         [Fact]
@@ -543,16 +142,13 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.SkipUncoveredMutants
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
-                // test 0 and 1 cover mutant 1
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = ";", ["T1"] = ";" }, endProcess);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
+            // test 0 and 1 cover mutant 1
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = ";", ["T1"] = ";" });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
-               _mutant.CoveringTests.Count.ShouldBe(0);
-            }
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
+            Mutant.CoveringTests.Count.ShouldBe(0);
         }
 
         [Fact]
@@ -563,24 +159,21 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
-                // only first test covers one mutant
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = ";" }, endProcess);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
+            // only first test covers one mutant
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = ";" });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
 
-                SetupMockPartialTestRun(mockVsTest, new Dictionary<string, string> { ["0"] = "T0=S" }, endProcess);
+            SetupMockPartialTestRun(mockVsTest, new Dictionary<string, string> { ["0"] = "T0=S" });
 
-                // process coverage information
-                var result = runner.TestMultipleMutants(null, new []{_mutant}, null);
-                // verify Abort has been called
-                Mock.Verify(mockVsTest);
-                // verify only one test has been run
-                result.RanTests.Count.ShouldBe(1);
-            }
+            // process coverage information
+            var result = runner.TestMultipleMutants(null, new[] { Mutant }, null);
+            // verify Abort has been called
+            Mock.Verify(mockVsTest);
+            // verify only one test has been run
+            result.RanTests.Count.ShouldBe(1);
         }
 
         [Fact]
@@ -591,26 +184,23 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
-                // only first test covers one mutant
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;0", ["T1"] = ";" }, endProcess);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
+            // only first test covers one mutant
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;0", ["T1"] = ";" });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
 
-                SetupMockTestRun(mockVsTest, false, endProcess, TestCases);
-                // mutant 0 is covered
-                _mutant.IsStaticValue.ShouldBeTrue();
-                var result = runner.TestMultipleMutants(null, new []{_mutant}, null);
-                // mutant is killed
-                result.FailingTests.IsEmpty.ShouldBeFalse();
-                // mutant 1 is not covered
-                 result = runner.TestMultipleMutants(null, new []{_otherMutant}, null);
-                // tests are ok
-                result.RanTests.IsEmpty.ShouldBeTrue();
-            }
+            SetupMockTestRun(mockVsTest, false, TestCases);
+            // mutant 0 is covered
+            Mutant.IsStaticValue.ShouldBeTrue();
+            var result = runner.TestMultipleMutants(null, new[] { Mutant }, null);
+            // mutant is killed
+            result.FailingTests.IsEmpty.ShouldBeFalse();
+            // mutant 1 is not covered
+            result = runner.TestMultipleMutants(null, new[] { OtherMutant }, null);
+            // tests are ok
+            result.RanTests.IsEmpty.ShouldBeTrue();
         }
 
         [Fact]
@@ -622,36 +212,21 @@ namespace Stryker.Core.UnitTest.TestRunners
                 Concurrency = Math.Max(Environment.ProcessorCount / 2, 1)
             };
 
-            var mutantFilter = new Mock<IMutantFilter>(MockBehavior.Loose);
+            var project = BuildTestProject(new []{Mutant, OtherMutant, new Mutant{Id = 2}, new Mutant{Id = 4}});
+            // make sure we have 4 mutants
+            var myTestCases = TestCases.ToList();
+            myTestCases.Add(BuildCase("T2"));
+            myTestCases.Add(BuildCase ("T3"));
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner, myTestCases);
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
+            var tester = BuildMutationTestProcess(runner, options, targetProject:project);
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = "1;" });
+            tester.GetCoverage();
+            SetupMockPartialTestRun(mockVsTest, new Dictionary<string, string> { ["0,1"] = "T0=S,T1=F" });
+            tester.Test(project.ProjectContents.Mutants.Where(x => x.ResultStatus == MutantStatus.NotRun));
 
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
-                // make sure we have 4 mutants
-                _projectContents.Add(new CsharpFileLeaf { Mutants = new[] { new Mutant { Id = 2 }, new Mutant { Id = 3 } } });
-                TestCases.Add(new TestCase("T2", _executorUri, _testAssemblyPath));
-                TestCases.Add(new TestCase("T3", _executorUri, _testAssemblyPath));
-
-                var ranTests = new TestsGuidList(TestCases.Select(t => t.Id));
-                var testRunResult = new TestRunResult(ranTests, TestsGuidList.NoTest(), TestsGuidList.NoTest(),
-                    string.Empty, TimeSpan.Zero);
-                var input = new MutationTestInput
-                {
-                    ProjectInfo = _targetProject,
-                    TestRunner = runner,
-                    InitialTestRun = new InitialTestRun(testRunResult, new TimeoutValueCalculator(500))
-                };
-                var mockReporter = new Mock<IReporter>();
-                var tester = new MutationTestProcess(input, mockReporter.Object, new MutationTestExecutor(input.TestRunner), fileSystem: _fileSystem, options: options, mutantFilter: mutantFilter.Object);
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;", ["T1"] = "1;" }, endProcess);
-                tester.GetCoverage();
-                SetupMockPartialTestRun(mockVsTest, new Dictionary<string, string> { ["1,0"] = "T0=S,T1=F" }, endProcess);
-                tester.Test(_projectContents.Mutants.Where(x => x.ResultStatus == MutantStatus.NotRun));
-
-                _mutant.ResultStatus.ShouldBe(MutantStatus.Survived);
-                _otherMutant.ResultStatus.ShouldBe(MutantStatus.Killed);
-            }
+            Mutant.ResultStatus.ShouldBe(MutantStatus.Survived);
+            OtherMutant.ResultStatus.ShouldBe(MutantStatus.Killed);
         }
 
         [Fact]
@@ -662,37 +237,33 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.CoverageBasedTest | OptimizationModes.CaptureCoveragePerTest
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
 
-                SetupMockCoveragePerTestRunP(mockVsTest, new Dictionary<string, string> { ["T0"] = "0,1;1", ["T1"] = ";" }, endProcess);
+            SetupMockCoveragePerTestRunP(mockVsTest, new Dictionary<string, string> { ["T0"] = "0,1;1", ["T1"] = ";" });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
 
-                SetupMockPartialTestRun(mockVsTest, new Dictionary<string, string> { ["0"] = "T0=F", ["1"] = "T0=S" }, endProcess);
-                var result = runner.TestMultipleMutants(null, new []{_otherMutant}, null);
-                result.FailingTests.IsEmpty.ShouldBeTrue();
-                result = runner.TestMultipleMutants(null, new []{_mutant}, null);
-                result.FailingTests.IsEmpty.ShouldBeFalse();
-            }
+            SetupMockPartialTestRun(mockVsTest, new Dictionary<string, string> { ["0"] = "T0=F", ["1"] = "T0=S" });
+            var result = runner.TestMultipleMutants(null, new[] { OtherMutant }, null);
+            result.FailingTests.IsEmpty.ShouldBeTrue();
+            result = runner.TestMultipleMutants(null, new[] { Mutant }, null);
+            result.FailingTests.IsEmpty.ShouldBeFalse();
         }
 
         [Fact]
         public void HandleMultipleTestResults()
         {
             var options = new StrykerOptions();
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
             // assume 2 results for T0
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) }, endProcess);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) });
             var result = runner.InitialTest();
             // initial test is fine
             result.FailingTests.IsEmpty.ShouldBeTrue();
             // test session will fail
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T0", true), ("T1", true) }, endProcess);
-            result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T0", true), ("T1", true) });
+            result = runner.TestMultipleMutants(null, new []{Mutant}, null);
             result.FailingTests.IsEmpty.ShouldBeTrue();
             result.RanTests.IsEveryTest.ShouldBeTrue();
         }
@@ -701,15 +272,14 @@ namespace Stryker.Core.UnitTest.TestRunners
         public void HandleMultipleTestResultsForXUnit()
         {
             var options = new StrykerOptions();
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
             var tests = new List<TestCase>
             {
                 new("X0", _xUnitUri, "test.cs"),
                 new("X1", _xUnitUri, "test.cs")
             };
-            var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner, tests);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner, tests);
             // assume 2 results for T0
-            SetupMockTestRun(mockVsTest, new[] { ("X0", true), ("X1", true), ("X0", true) , ("X0", true), ("X0", true)}, endProcess, tests);
+            SetupMockTestRun(mockVsTest, new[] { ("X0", true), ("X1", true), ("X0", true) , ("X0", true), ("X0", true)}, tests);
             var result = runner.InitialTest();
             // the duration should be less than 3 times the (test) default duration
             result.Duration.ShouldBeLessThan(_testDefaultDuration.Duration()*3);
@@ -719,23 +289,22 @@ namespace Stryker.Core.UnitTest.TestRunners
         public void HandleFailureWithinMultipleTestResults()
         {
             var options = new StrykerOptions();
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
             // assume 2 results for T0
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) }, endProcess);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) });
             var result = runner.InitialTest();
             // initial test is fine
             result.FailingTests.IsEmpty.ShouldBeTrue();
             // test session will fail on test 1
-            SetupMockTestRun(mockVsTest, new[] { ("T0", false), ("T0", true), ("T1", true) }, endProcess);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", false), ("T0", true), ("T1", true) });
 
-            result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            result = runner.TestMultipleMutants(null, new []{Mutant}, null);
             result.RanTests.IsEveryTest.ShouldBeTrue();
             result.FailingTests.IsEmpty.ShouldBeFalse();
             result.FailingTests.GetGuids().ShouldContain(TestCases[0].Id);
             // test session will fail on the other test result
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T0", false), ("T1", true) }, endProcess);
-            result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T0", false), ("T1", true) });
+            result = runner.TestMultipleMutants(null, new []{Mutant}, null);
             result.RanTests.IsEveryTest.ShouldBeTrue();
             result.FailingTests.IsEmpty.ShouldBeFalse();
             result.FailingTests.GetGuids().ShouldContain(TestCases[0].Id);
@@ -745,16 +314,15 @@ namespace Stryker.Core.UnitTest.TestRunners
         public void HandleTimeOutWithMultipleTestResults()
         {
             var options = new StrykerOptions();
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
             // assume 2 results for T0
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) }, endProcess);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) });
             var result = runner.InitialTest();
             // initial test is fine
             result.FailingTests.IsEmpty.ShouldBeTrue();
             // test session will fail
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true) }, endProcess);
-            result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true) });
+            result = runner.TestMultipleMutants(null, new []{Mutant}, null);
 
             result.FailingTests.IsEmpty.ShouldBeTrue();
             result.TimedOutTests.Count.ShouldBe(1);
@@ -766,16 +334,15 @@ namespace Stryker.Core.UnitTest.TestRunners
         public void HandleFailureWhenExtraMultipleTestResults()
         {
             var options = new StrykerOptions();
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
             // assume 2 results for T0
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) }, endProcess);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) });
             var result = runner.InitialTest();
             // initial test is fine
             result.FailingTests.IsEmpty.ShouldBeTrue();
             // test session will fail
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T0", true), ("T0", false), ("T1", true) }, endProcess);
-            result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T0", true), ("T0", false), ("T1", true) });
+            result = runner.TestMultipleMutants(null, new []{Mutant}, null);
             result.RanTests.IsEveryTest.ShouldBeTrue();
             result.FailingTests.IsEmpty.ShouldBeFalse();
             result.FailingTests.GetGuids().ShouldContain(TestCases[0].Id);
@@ -785,16 +352,15 @@ namespace Stryker.Core.UnitTest.TestRunners
         public void HandleUnexpectedTestResult()
         {
             var options = new StrykerOptions();
-            using var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset);
-            var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
             // assume 2 results for T0
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) }, endProcess);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T1", true), ("T0", true) });
             var result = runner.InitialTest();
             // initial test is fine
             result.FailingTests.IsEmpty.ShouldBeTrue();
             // test session will fail
-            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T2", true), ("T1", true), ("T0", true) }, endProcess);
-            result = runner.TestMultipleMutants(null, new []{_mutant}, null);
+            SetupMockTestRun(mockVsTest, new[] { ("T0", true), ("T2", true), ("T1", true), ("T0", true) });
+            result = runner.TestMultipleMutants(null, new []{Mutant}, null);
             result.RanTests.IsEveryTest.ShouldBeTrue();
             result.FailingTests.IsEmpty.ShouldBeTrue();
         }
@@ -809,17 +375,14 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
 
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;|1", ["T1"] = ";"}, endProcess);
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "0;|1", ["T1"] = ";" });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
-                // the suspicious mutant should be tested against all tests
-                _otherMutant.CoveringTests.Count.ShouldBe(2);
-            }
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
+            // the suspicious mutant should be tested against all tests
+            OtherMutant.CoveringTests.Count.ShouldBe(2);
         }
 
         // this verifies that tests missing any coverage information are
@@ -832,45 +395,17 @@ namespace Stryker.Core.UnitTest.TestRunners
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            using (var endProcess = new EventWaitHandle(false, EventResetMode.ManualReset))
-            {
-                var mockVsTest = BuildVsTestRunnerPool(options, endProcess, out var runner);
+            var mockVsTest = BuildVsTestRunnerPool(options, out var runner);
 
-                SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "1;", ["T1"] = null}, endProcess);
+            SetupMockCoverageRun(mockVsTest, new Dictionary<string, string> { ["T0"] = "1;", ["T1"] = null });
 
-                var analyzer = new CoverageAnalyser(options);
-                analyzer.DetermineTestCoverage(runner, new []{_mutant, _otherMutant});
-                // the suspicious mutant should be tested against all tests
-                _otherMutant.CoveringTests.Count.ShouldBe(2);
-                _mutant.CoveringTests.Count.ShouldBe(1);
-            }
+            var analyzer = new CoverageAnalyser(options);
+            analyzer.DetermineTestCoverage(runner, new[] { Mutant, OtherMutant });
+            // the suspicious mutant should be tested against all tests
+            OtherMutant.CoveringTests.Count.ShouldBe(2);
+            Mutant.CoveringTests.Count.ShouldBe(1);
         }
 
         // class mocking the VsTest Host Launcher
-        private class MoqHost : IStrykerTestHostLauncher
-        {
-            private readonly WaitHandle _handle;
-
-            public MoqHost(WaitHandle handle, bool succeed, bool isDebug)
-            {
-                _handle = handle;
-                IsProcessCreated = succeed;
-                IsDebug = isDebug;
-            }
-
-            public int LaunchTestHost(TestProcessStartInfo defaultTestHostStartInfo)
-            {
-                throw new NotImplementedException();
-            }
-
-            public int LaunchTestHost(TestProcessStartInfo defaultTestHostStartInfo, CancellationToken cancellationToken)
-            {
-                throw new NotImplementedException();
-            }
-
-            public bool IsDebug { get; }
-
-            public bool IsProcessCreated { get; }
-        }
     }
 }
